+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Format based on [Keep a Changelog](https://keepachangelog.com).
 
-## [Unreleased]
+## [0.3.1] — 2026-05-17
 
 ### Changed